# backend/public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaya Test",
        "version": "0.1"
    },
    "paths": {
        "/api/payment": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Make a payment",
                "operationId": "82f805bcf2dd8149b522bead8ae17044",
                "requestBody": {
                    "description": "Dados do pagamento",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "245.90"
                                    },
                                    "installments": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "ae4e50b2a8f3h6d9f2c3a4b5d6e7f8g9"
                                    },
                                    "payment_method_id": {
                                        "type": "string",
                                        "example": "master"
                                    },
                                    "payer": {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "example_random@gmail.com"
                                            },
                                            "identification": {
                                                "properties": {
                                                    "type": {
                                                        "type": "string",
                                                        "example": "CPF"
                                                    },
                                                    "number": {
                                                        "type": "string",
                                                        "example": "12345678909"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "eda2d5f0-dbfe-4093-84c2-55ab47febe0f"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-02-25"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": "eda2d5f0-dbfe-4093-84c2-55ab47febe0f",
                                        "created_at": "2024-02-25"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request without payload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not provided in the request body"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Payment not provided in the request body"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request with invalid field or value",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid payment provided.The possible reasons are:A field of the provided payment was null or with invalid values"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Invalid payment provided.The possible reasons are:A field of the provided payment was null or with invalid values"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/rest/payments/{id}}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "List a payment by ID",
                "operationId": "331a4fb40378b7e76a0d7f48b65ac60e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso com itens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "10763b2b-5034-4cc0-8cb2-c5926439158c"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "transaction_amount": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 245.9
                                            },
                                            "installments": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "token": {
                                                "type": "string",
                                                "example": "ae4e50b2a8f3h6d9f2c3a4b5d6e7f8g9"
                                            },
                                            "payment_method_id": {
                                                "type": "string",
                                                "example": "master"
                                            },
                                            "payer": {
                                                "properties": {
                                                    "entity_type": {
                                                        "type": "string",
                                                        "example": "individual"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "customer"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example_random@gmail.com"
                                                    },
                                                    "identification": {
                                                        "properties": {
                                                            "type": {
                                                                "type": "string",
                                                                "example": "CPF"
                                                            },
                                                            "number": {
                                                                "type": "string",
                                                                "example": "12345678909"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "notification_url": {
                                                "type": "string",
                                                "example": "https://webhook.site/e3d32ab9-737a-4832-a5fa-f36a172cec53"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-02-25"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found with the specified id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rest/payments/{id}": {
            "delete": {
                "tags": [
                    "Payment"
                ],
                "summary": "Cancel a payment by ID",
                "operationId": "160d98f03e8f2ebf2c1d18c1930bea43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "CANCELED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank slip not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Payment not found with the specified id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Payment"
                ],
                "summary": "Confirm a payment by ID",
                "operationId": "0bd670abb4f45006e0bf4e89e15440bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the payment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "PAID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bank slip not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Bank slip not found with the specified id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}